Cleanup debug print statements and update tests
    - create helper functions for client wrapper tests
    - move server up calls to seperate file for use in
      both sched.py and server.py, so code does not
      have to be duplicated in both files
    - cleaned up print statements in pmix.pyx and
      pmix.pxi
